# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube Channel Feed API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Channel {
  id: number;
  youtubeChannelId: string;
  name: string;
  thumbnailUrl?: string | null;
  addedAt: string;
}

export interface AddChannelRequest {
  /** YouTube channel ID (starts with UC...) or handle (@username) */
  youtubeChannelId: string;
}

export interface Video {
  videoId: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  publishedAt: string;
  viewCount?: string | null;
  channelId: string;
  channelName: string;
  channelThumbnailUrl?: string | null;
  duration?: string | null;
}

export interface ChannelSearchResult {
  youtubeChannelId: string;
  name: string;
  description: string;
  thumbnailUrl?: string | null;
  subscriberCount?: string | null;
}

export interface VideoSummaryRequest {
  videoId: string;
  title: string;
  description: string;
  channelName: string;
}

export interface VideoSummaryResponse {
  summary: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface RecordingItem {
  id: number;
  beatVideoId: string;
  beatTitle: string;
  beatChannelName: string;
  beatThumbnailUrl: string;
  objectPath: string;
  durationSeconds: number;
  createdAt: string;
}

export interface CreateRecordingBody {
  beatVideoId: string;
  beatTitle: string;
  beatChannelName: string;
  beatThumbnailUrl: string;
  objectPath: string;
  durationSeconds: number;
}

export interface DawLane {
  id: number;
  name: string;
  color: string;
  muted: boolean;
  volume: number;
  startOffset: number;
  durationSec: number;
  objectPath?: string | null;
  mime: string;
}

export interface DawProjectItem {
  id: number;
  name: string;
  beatVideoId: string;
  beatTitle: string;
  beatChannelName: string;
  beatThumbnailUrl: string;
  lanes: DawLane[];
  createdAt: string;
  updatedAt: string;
}

export interface YogaVideo {
  id: number;
  videoId: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  channelId: string;
  channelName: string;
  channelThumbnailUrl?: string | null;
  viewCount?: string | null;
  duration?: string | null;
  publishedAt: string;
  category?: string | null;
  savedAt: string;
}

export interface AddYogaVideoBody {
  url: string;
  category?: string | null;
}

export interface UpdateYogaCategoryBody {
  category?: string | null;
}

export interface SavedVideoItem {
  id: number;
  videoId: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  channelId: string;
  channelName: string;
  channelThumbnailUrl?: string | null;
  viewCount?: string | null;
  duration?: string | null;
  publishedAt: string;
  savedAt: string;
  mp3Ready: boolean;
}

export interface SaveVideoBody {
  url: string;
}

export interface SavedVideoStatus {
  saved: boolean;
}

export interface CreateDawProjectBody {
  name: string;
  beatVideoId: string;
  beatTitle: string;
  beatChannelName: string;
  beatThumbnailUrl: string;
  lanes: DawLane[];
}

export type SearchChannelsParams = {
  /**
   * Search query
   */
  q: string;
};

export type ListVideosParams = {
  /**
   * Filter by specific channel DB id
   */
  channelId?: number;
};

export type SearchBeatChannelsParams = {
  q: string;
};

export type ListBeatsParams = {
  channelId?: number;
};

export type SearchBeatsParams = {
  /**
   * Search query
   */
  q: string;
  /**
   * Sort order
   */
  order?: SearchBeatsOrder;
  maxResults?: number;
};

export type SearchBeatsOrder =
  (typeof SearchBeatsOrder)[keyof typeof SearchBeatsOrder];

export const SearchBeatsOrder = {
  relevance: "relevance",
  date: "date",
  viewCount: "viewCount",
} as const;

export type GetSimilarBeatsParams = {
  title: string;
};

export type ListYogaVideosParams = {
  category?: string;
};