# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YouTube Channel Feed API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saved:
    get:
      operationId: listSavedVideos
      tags: [recordings]
      summary: List all saved videos
      responses:
        "200":
          description: List of saved videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedVideoItem"
    post:
      operationId: saveVideo
      tags: [recordings]
      summary: Save a video by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveVideoBody"
      responses:
        "201":
          description: Saved video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedVideoItem"
  /saved/{videoId}:
    get:
      operationId: checkSavedVideo
      tags: [recordings]
      summary: Check if a video is saved
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Saved status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedVideoStatus"
        "404":
          description: Not saved
    delete:
      operationId: removeSavedVideo
      tags: [recordings]
      summary: Remove a saved video
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SavedVideoItem:
      type: object
      properties:
        id:
          type: integer
        videoId:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        channelId:
          type: string
        channelName:
          type: string
        channelThumbnailUrl:
          type: string
          nullable: true
        viewCount:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        publishedAt:
          type: string
        savedAt:
          type: string
        mp3Ready:
          type: boolean
      required:
        - id
        - videoId
        - title
        - description
        - thumbnailUrl
        - channelId
        - channelName
        - publishedAt
        - savedAt
        - mp3Ready
    SaveVideoBody:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    SavedVideoStatus:
      type: object
      properties:
        saved:
          type: boolean
      required:
        - saved
...