# lib/api-zod/src/generated/types/savedVideoItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube Channel Feed API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedVideoItem {
  id: number;
  videoId: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  channelId: string;
  channelName: string;
  channelThumbnailUrl?: string | null;
  viewCount?: string | null;
  duration?: string | null;
  publishedAt: string;
  savedAt: string;
  mp3Ready: boolean;
}